Module burn_core::backend

Backend module.

Re-exports

pub use burn_ndarray as ndarray;

pub use burn_autodiff as autodiff;

pub use burn_wgpu as wgpu;

pub use burn_tch as libtorch;

Structs

Autodiff

Enable auto-differentiation on a backend.

Fusion

Enable dynamic operation fusion on a backend that implements fusion backend.

LibTorch

Tensor backend that uses LibTorch with the [tch] crate for executing tensor operations.

NdArray

Tensor backend that uses the ndarray crate for executing tensor operations.

Wgpu

Tensor backend that uses the [wgpu] crate for executing GPU compute shaders.