Trait burn_core::lr_scheduler::LrScheduler

pub trait LrScheduler<B: Backend>: Send + Sync {
    type Record: Record<B>;

    // Required methods
    fn step(&mut self) -> LearningRate;
    fn to_record(&self) -> Self::Record;
    fn load_record(self, record: Self::Record) -> Self;
}

Learning rate scheduler defines how the learning rate will evolve during training.

Required Associated Types

type Record: Record<B>

Scheduler associative type to be used when saving and loading the state.

Required Methods

fn step(&mut self) -> LearningRate

Perform the scheduler step, potentially updating its state, and returning the effective learning rate.

fn to_record(&self) -> Self::Record

Get the current state of the scheduler as a record.

fn load_record(self, record: Self::Record) -> Self

Load the state of the scheduler as a record.

Object Safety

This trait is not object safe.