burn_core

Module module

Module for the neural network module.

Macros

constant

Constant macro.

Structs

Attribute

Attribute to print in the display method.

ConstantRecord

Record used for constant type implementing the module trait.

Content

Struct to store the attributes of a module for formatting.

DisplaySettings

Custom module display settings.

Ignored

Container to satisfy the Module trait for types that are not modules.

Param

Parameters are the fundamental building blocks of modules where they serve as containers for tensors that can be updated during training, and loaded during inference. If you don’t want to save the tensors with a record and/or don’t want to update it during training, you don’t need this type to wrap your tensor.

ParamId

Parameter ID.

Quantizer

Describes how to quantize a module.

RunningState

A state that can be updated during the forward pass while being thread safe.

Traits

AutodiffModule

Module with auto-differentiation backend.

Module

Trait for all neural network modules.

ModuleDisplay

Trait to implement custom display settings for a module.

ModuleDisplayDefault

Default display settings for a module.

ModuleMapper

Module mapper trait.

ModuleVisitor

Module visitor trait.

Parameter

Trait that defines what is necessary for a type to be a parameter.

Functions

extract_type_name

Extracts the short name of a type T

list_param_ids

List all the parameter ids in a module.

Type Aliases

Devices

Type alias to Vec<B::Device> which supports no_std environments, but automatically using the alloc crate.

Derive Macros

Module

Derive macro for the module.