burn_core::module

Struct Param

pub struct Param<T: Parameter> {
    pub id: ParamId,
    /* private fields */
}

Parameters are the fundamental building blocks of modules where they serve as containers for tensors that can be updated during training, and loaded during inference. If you don’t want to save the tensors with a record and/or don’t want to update it during training, you don’t need this type to wrap your tensor.

Laziness

The initialization of parameters can be lazy when created using uninitialized, which can be done using an initializer.

This reduces the amount of allocations done when loading a model for inference without having to create a custom initialization function only for inference.

Example

let device = Device::default();
let config = ModuleConfig::default();
let record = Recorder::new().load("/path/to/module", &device);

// No tensor allocation
let module = config.init(device);
// Will use the tensor allocated for the record if the same device is used.
let module = module.load_record(record);

Fields
id: ParamId

The unique ID of this parameter. This is used by eg. optimizers to associate a gradient with a specific parameter.

Implementations

impl<T: Parameter> Param<T>

pub fn initialized(id: ParamId, value: T) -> Self

Create a new parameter that is already initialized.

pub fn uninitialized<F>( id: ParamId, init: F, device: T::Device, is_require_grad: bool, ) -> Self
where F: FnOnce(&T::Device, bool) -> T + Send + 'static,

Create a new parameter that is not already initialized.

pub fn val(&self) -> T

Gets the parameter value.

pub fn into_value(self) -> T

Gets the parameter’s value while consuming the parameter.

pub fn consume(self) -> (ParamId, T)

Gets the parameter id and value while consuming the parameter.

pub fn map<F: FnOnce(T) -> T>(self, func: F) -> Self

Execute the given function on the inner value.

pub fn set_require_grad(self, require_grad: bool) -> Self

Override the gradient requirement for the current parameter.

impl<B: Backend, const D: usize> Param<Tensor<B, D>>

pub fn from_tensor(value: Tensor<B, D>) -> Self

Create a new parameter from a float tensor.

Warnings

We strongly recommend using Param::uninitialized if you are using this method to initialize parameters inside a module, since the tensor initialization will be lazy, making the loading of weights more performant.

pub fn from_data<T>(data: T, device: &B::Device) -> Self
where T: Into<TensorData>,

Create a new parameter from data.