Trait burn_core::module::Module

pub trait Module<B: Backend>: Clone + Send + Sync + Debug {
    type Record: Record;

    // Required methods
    fn visit<V: ModuleVisitor<B>>(&self, visitor: &mut V);
    fn map<M: ModuleMapper<B>>(self, mapper: &mut M) -> Self;
    fn load_record(self, record: Self::Record) -> Self;
    fn into_record(self) -> Self::Record;

    // Provided methods
    fn devices(&self) -> Vec<B::Device> { ... }
    fn fork(self, device: &B::Device) -> Self { ... }
    fn to_device(self, device: &B::Device) -> Self { ... }
    fn no_grad(self) -> Self { ... }
    fn num_params(&self) -> usize { ... }
    fn save_file<FR: FileRecorder, PB: Into<PathBuf>>(
        self,
        file_path: PB,
        recorder: &FR
    ) -> Result<(), RecorderError> { ... }
    fn load_file<FR: FileRecorder, PB: Into<PathBuf>>(
        self,
        file_path: PB,
        recorder: &FR
    ) -> Result<Self, RecorderError> { ... }
}

Trait for all neural network modules.

Modules should be created using the derive attribute. This will make your module trainable, savable and loadable via state and load.

Example

A module should have a backend defined as a generic parameter B. This will be used by the derive attribute to generate the code necessary to optimize and train the module on any backend.

// Not necessary when using the burn crate directly.
use burn_core as burn;

use burn::{
    nn,
    module::Module,
    tensor::Tensor,
    tensor::backend::Backend,
};

#[derive(Module, Debug)]
struct MyModule<B: Backend> {
  my_param: nn::Linear<B>,
  my_other_field: usize,
}

Required Associated Types

type Record: Record

Type to save and load the module.

Required Methods

fn visit<V: ModuleVisitor<B>>(&self, visitor: &mut V)

Visit each tensor in the module with a visitor.

fn map<M: ModuleMapper<B>>(self, mapper: &mut M) -> Self

Map each tensor in the module with a mapper.

fn load_record(self, record: Self::Record) -> Self

Load the module state from a record.

fn into_record(self) -> Self::Record

Convert the module into a record containing the state.

Provided Methods

fn devices(&self) -> Vec<B::Device>

Get the device list of the module and all of its sub-modules.

fn fork(self, device: &B::Device) -> Self

Fork the module and all of its sub-modules to the given device.

Notes

This is similar to to_device, but it ensures the module will have its own autodiff graph.

fn to_device(self, device: &B::Device) -> Self

Move the module and all of its sub-modules to the given device.

Warnings

The device operations will be registered in the autodiff graph. Therefore, be sure to call backward only one time even if you have the same module on multiple devices. If you want to call backward multiple times, look into using fork instead.

fn no_grad(self) -> Self

Each tensor in the module tree will not require grad.

Warnings

This should not be used for inference, use valid when using AD modules. This is mostly useful when performing partial finetuning, which is updating only a small fraction of the parameters instead of finetuning all of them.

fn num_params(&self) -> usize

Get the number of parameters the module has, including all of its sub-modules.

fn save_file<FR: FileRecorder, PB: Into<PathBuf>>( self, file_path: PB, recorder: &FR ) -> Result<(), RecorderError>

Save the module to a file using the provided file recorder.

List of supported file recorders:

default

bincode

bincode compressed with gzip

json pretty

json compressed with gzip

named mpk

named mpk compressed with gzip

Notes

The file extension is automatically added depending on the file recorder provided, you don’t have to specify it.

fn load_file<FR: FileRecorder, PB: Into<PathBuf>>( self, file_path: PB, recorder: &FR ) -> Result<Self, RecorderError>

Load the module from a file using the provided file recorder.

The recorder should be the same as the one used to save the module, see save_file.

Notes

The file extension is automatically added depending on the file recorder provided, you don’t have to specify it.