Struct burn_core::nn::Embedding

pub struct Embedding<B: Backend> {
    pub weight: Param<Tensor<B, 2>>,
}

Lookup table to store a fix number of vectors.

Params

weight: Matrix of shape [n_embedding, d_model] initialized from a normal distribution: N(0, 1)

Fields
weight: Param<Tensor<B, 2>>

The learnable weights of the module of shape [n_embedding, d_model] initialized from a normal distribution N(0, 1).

Implementations

impl<B: Backend> Embedding<B>

pub fn forward(&self, input: Tensor<B, 2, Int>) -> Tensor<B, 3>

Applies the forward pass on the input tensor.

Shapes

input: [batch_size, seq_length]

output: [batch_size, d_model]