Struct burn_core::nn::PositionalEncoding

pub struct PositionalEncoding<B: Backend> { /* private fields */ }

Positional encoding layer for transformer models.

This layer adds positional information to the input embeddings, allowing the transformer model to take into account the order of the sequence. The positional encoding is added to the input embeddings by computing a set of sinusoidal functions with different frequencies and phases.

Sinusoids are used for positional embedding introduced in Attention is all you need.

The reference implementation can be found here: LANGUAGE MODELING WITH NN.TRANSFORMER AND TORCHTEXT

Implementations

impl<B: Backend> PositionalEncoding<B>

pub fn forward(&self, input: Tensor<B, 3>) -> Tensor<B, 3>

Applies the forward pass on the input tensor by adding the sinusoids to the input.

Shapes

input: [batch_size, seq_length, d_model]

output: [batch_size, seq_length, d_model]

Panics

Panics if the input sequence length is greater than the maximum sequence size.

Panics if the input d_model is not equal to the d_model of the sinusoids.