Trait burn_core::optim::Optimizer

pub trait Optimizer<M, B>: Send + Syncwhere
    M: AutodiffModule<B>,
    B: AutodiffBackend,{
    type Record: Record;

    // Required methods
    fn step(&mut self, lr: LearningRate, module: M, grads: GradientsParams) -> M;
    fn to_record(&self) -> Self::Record;
    fn load_record(self, record: Self::Record) -> Self;
}

General trait to optimize module.

Required Associated Types

type Record: Record

Optimizer associative type to be used when saving and loading the state.

Required Methods

fn step(&mut self, lr: LearningRate, module: M, grads: GradientsParams) -> M

Perform the optimizer step using the given learning rate and gradients. The updated module is returned.

fn to_record(&self) -> Self::Record

Get the current state of the optimizer as a record.

fn load_record(self, record: Self::Record) -> Self

Load the state of the optimizer as a record.