burn_core::prelude

Struct TensorData

pub struct TensorData {
    pub shape: Vec<usize>,
    pub dtype: DType,
    /* private fields */
}

Data structure for tensors.

Fields
shape: Vec<usize>

The shape of the tensor.

dtype: DType

The data type of the tensor.

Implementations

impl TensorData

pub fn new<E, S>(value: Vec<E>, shape: S) -> TensorData
where E: Element, S: Into<Vec<usize>>,

Creates a new tensor data structure.

pub fn quantized<E, S>( value: Vec<E>, shape: S, strategy: QuantizationStrategy, ) -> TensorData
where E: Element, S: Into<Vec<usize>>,

Creates a new quantized tensor data structure.

pub fn from_bytes<S>(bytes: Vec<u8>, shape: S, dtype: DType) -> TensorData
where S: Into<Vec<usize>>,

Creates a new tensor data structure from raw bytes.

Prefer TensorData::new or TensorData::quantized over this method unless you are certain that the bytes representation is valid.

pub fn as_slice<E>(&self) -> Result<&[E], DataError>
where E: Element,

Returns the immutable slice view of the tensor data.

pub fn as_mut_slice<E>(&mut self) -> Result<&mut [E], DataError>
where E: Element,

Returns the mutable slice view of the tensor data.

Panics

If the target element type is different from the stored element type.

pub fn to_vec<E>(&self) -> Result<Vec<E>, DataError>
where E: Element,

Returns the tensor data as a vector of scalar values.

pub fn into_vec<E>(self) -> Result<Vec<E>, DataError>
where E: Element,

Returns the tensor data as a vector of scalar values.

pub fn iter<E>(&self) -> Box<dyn Iterator<Item = E> + '_>
where E: Element,

Returns an iterator over the values of the tensor data.

pub fn num_elements(&self) -> usize

Returns the total number of elements of the tensor data.

pub fn random<E, R, S>( shape: S, distribution: Distribution, rng: &mut R, ) -> TensorData
where E: Element, R: RngCore, S: Into<Vec<usize>>,

Populates the data with random values.

pub fn zeros<E, S>(shape: S) -> TensorData
where E: Element, S: Into<Vec<usize>>,

Populates the data with zeros.

pub fn ones<E, S>(shape: S) -> TensorData
where E: Element, S: Into<Vec<usize>>,

Populates the data with ones.

pub fn full<E, S>(shape: S, fill_value: E) -> TensorData
where E: Element, S: Into<Vec<usize>>,

Populates the data with the given value

pub fn convert<E>(self) -> TensorData
where E: Element,

Converts the data to a different element type.

pub fn as_bytes(&self) -> &[u8]

Returns the data as a slice of bytes.

pub fn into_bytes(self) -> Bytes

Returns the bytes representation of the data.

pub fn with_quantization(self, quantization: QuantizationStrategy) -> TensorData

Applies the data quantization strategy.

Panics

Panics if the data type is not supported for quantization.

pub fn dequantize(self) -> Result<TensorData, DataError>

Dequantizes the data according to its quantization scheme.

pub fn assert_approx_eq(&self, other: &TensorData, precision: usize)

Asserts the data is approximately equal to another data.

Arguments

other - The other data.

precision - The precision of the comparison.

Panics

Panics if the data is not approximately equal.

pub fn assert_eq(&self, other: &TensorData, strict: bool)

Asserts the data is equal to another data.

Arguments

other - The other data.

strict - If true, the data types must the be same. Otherwise, the comparison is done in the current data type.

Panics

Panics if the data is not equal.

pub fn assert_approx_eq_diff(&self, other: &TensorData, tolerance: f64)

Asserts the data is approximately equal to another data.

Arguments

other - The other data.

tolerance - The tolerance of the comparison.

Panics

Panics if the data is not approximately equal.

pub fn assert_within_range<E>(&self, range: Range<E>)
where E: Element,

Asserts each value is within a given range.

Arguments

range - The range.

Panics

If any value is not within the half-open range bounded inclusively below and exclusively above (start..end).

pub fn assert_within_range_inclusive<E>(&self, range: RangeInclusive<E>)
where E: Element,

Asserts each value is within a given inclusive range.

Arguments

range - The range.

Panics

If any value is not within the half-open range bounded inclusively (start..=end).