burn_core

Module record

Module for the recorder.

Structs

BinBytesRecorder

In memory recorder using the bincode format.

BurnMetadata

Metadata of a record.

BurnRecord

Record that can be saved by a Recorder.

BurnRecordNoItem

Record that can be saved by a Recorder without the item.

DoublePrecisionSettings

Precision settings optimized for precision.

FullPrecisionSettings

Default precision settings.

HalfPrecisionSettings

Precision settings optimized for compactness.

ParamSerde

(De)serialize parameters into a clean format.

Enums

RecorderError

Error that can occur when using a Recorder.

Traits

BytesRecorder

Recorder trait specialized to save and load data to and from bytes.

PrecisionSettings

Settings allowing to control the precision when (de)serializing items.

Record

Trait to define a family of types which can be recorded using any settings.

Recorder

Record any item implementing Serialize and DeserializeOwned.

Type Aliases

NoStdInferenceRecorder

Inference recorder compatible with no-std.

Derive Macros

Record

Derive macro for the record.