Trait burn_core::record::serde::adapter::BurnModuleAdapter

pub trait BurnModuleAdapter: Sized {
    // Provided methods
    fn adapt(name: &str, data: NestedValue) -> NestedValue { ... }
    fn adapt_linear(data: NestedValue) -> NestedValue { ... }
    fn adapt_conv1d(data: NestedValue) -> NestedValue { ... }
    fn adapt_conv2d(data: NestedValue) -> NestedValue { ... }
    fn adapt_conv_transpose_1d(data: NestedValue) -> NestedValue { ... }
    fn adapt_conv_transpose_2d(data: NestedValue) -> NestedValue { ... }
    fn adapt_embedding(data: NestedValue) -> NestedValue { ... }
    fn adapt_group_norm(data: NestedValue) -> NestedValue { ... }
    fn adapt_layer_norm(data: NestedValue) -> NestedValue { ... }
    fn adapt_batch_norm(data: NestedValue) -> NestedValue { ... }
}

A trait that defines the adapter for a Burn module.

This is used to adapt an incoming module to a Burn module.

Provided Methods

fn adapt(name: &str, data: NestedValue) -> NestedValue

Adapts a module.

fn adapt_linear(data: NestedValue) -> NestedValue

Adapts a linear module.

fn adapt_conv1d(data: NestedValue) -> NestedValue

Adapts a Convolution 1D module.

fn adapt_conv2d(data: NestedValue) -> NestedValue

Adapts a Convolution 2D module.

fn adapt_conv_transpose_1d(data: NestedValue) -> NestedValue

Adapts convolution transpose 1D module.

fn adapt_conv_transpose_2d(data: NestedValue) -> NestedValue

Adapts convolution transpose 2D module.

fn adapt_embedding(data: NestedValue) -> NestedValue

Adapts embedding module.

fn adapt_group_norm(data: NestedValue) -> NestedValue

Adapts group normalization module.

fn adapt_layer_norm(data: NestedValue) -> NestedValue

Adapts layer normalization module.

fn adapt_batch_norm(data: NestedValue) -> NestedValue

Adapts batch normalization module.

Object Safety

This trait is not object safe.