Trait burn_core::record::Record

pub trait Record: Send + Sync {
    type Item<S: PrecisionSettings>: Serialize + DeserializeOwned;

    // Required methods
    fn into_item<S: PrecisionSettings>(self) -> Self::Item<S>;
    fn from_item<S: PrecisionSettings>(item: Self::Item<S>) -> Self;
}

Trait to define a family of types which can be recorded using any settings.

Required Associated Types

type Item<S: PrecisionSettings>: Serialize + DeserializeOwned

Type of the item that can be serialized and deserialized.

Required Methods

fn into_item<S: PrecisionSettings>(self) -> Self::Item<S>

Convert the current record into the corresponding item that follows the given settings.

fn from_item<S: PrecisionSettings>(item: Self::Item<S>) -> Self

Convert the given item into a record.