Trait burn_core::tensor::backend::ADBackend

pub trait ADBackend: Backend {
    type InnerBackend: Backend<Device = Self::Device, FloatElem = Self::FloatElem>;
    type Gradients: Send + Sync;

    // Required methods
    fn backward<const D: usize>(
        tensor: Self::TensorPrimitive<D>
    ) -> Self::Gradients;
    fn grad<const D: usize>(
        tensor: &Self::TensorPrimitive<D>,
        grads: &Self::Gradients
    ) -> Option<<Self::InnerBackend as Backend>::TensorPrimitive<D>>;
    fn grad_remove<const D: usize>(
        tensor: &Self::TensorPrimitive<D>,
        grads: &mut Self::Gradients
    ) -> Option<<Self::InnerBackend as Backend>::TensorPrimitive<D>>;
    fn grad_replace<const D: usize>(
        tensor: &Self::TensorPrimitive<D>,
        grads: &mut Self::Gradients,
        grad: <Self::InnerBackend as Backend>::TensorPrimitive<D>
    );
    fn inner<const D: usize>(
        tensor: Self::TensorPrimitive<D>
    ) -> <Self::InnerBackend as Backend>::TensorPrimitive<D>;
    fn from_inner<const D: usize>(
        tensor: <Self::InnerBackend as Backend>::TensorPrimitive<D>
    ) -> Self::TensorPrimitive<D>;
}

Trait that allows a backend to support autodiff.

Required Associated Types

type InnerBackend: Backend<Device = Self::Device, FloatElem = Self::FloatElem>

The inner backend type.

type Gradients: Send + Sync

Gradients type.

Required Methods

fn backward<const D: usize>(tensor: Self::TensorPrimitive<D>) -> Self::Gradients

Backward pass.

Arguments

tensor - The tensor is the last node of computational graph where the gradients are computed.

Returns

The gradients.

fn grad<const D: usize>( tensor: &Self::TensorPrimitive<D>, grads: &Self::Gradients ) -> Option<<Self::InnerBackend as Backend>::TensorPrimitive<D>>

Returns the gradients of a tensor.

Arguments

tensor - The tensor to extract the gradients from.

Returns

An optional tensor containing the gradient.

fn grad_remove<const D: usize>( tensor: &Self::TensorPrimitive<D>, grads: &mut Self::Gradients ) -> Option<<Self::InnerBackend as Backend>::TensorPrimitive<D>>

Pops the gradients of a tensor and returns them.

Arguments

tensor - The tensor to pop the gradients from.

grads - The gradients.

Returns

An optional tensor containing the given gradients.

fn grad_replace<const D: usize>( tensor: &Self::TensorPrimitive<D>, grads: &mut Self::Gradients, grad: <Self::InnerBackend as Backend>::TensorPrimitive<D> )

Replace the gradients of a tensor with the one provided.

If no gradient existed for the provided tensor, register it.

Arguments

tensor - The tensor to pop the gradients from.

grads - The gradients.

grad - The updated grad tensor.

fn inner<const D: usize>( tensor: Self::TensorPrimitive<D> ) -> <Self::InnerBackend as Backend>::TensorPrimitive<D>

Returns the tensor with inner backend type.

Arguments

tensor - The tensor to get the inner backend tensor for.

Returns

The inner backend tensor.

fn from_inner<const D: usize>( tensor: <Self::InnerBackend as Backend>::TensorPrimitive<D> ) -> Self::TensorPrimitive<D>

Converts the inner backend tensor to the autodiff backend tensor.

Arguments

tensor - The inner backend tensor to convert.

Returns

The autodiff backend tensor.