burn_core::tensor

Module ops

Operations on tensors module.

Modules

conv

Module with convolution operations.

pool

Module with pooling operations.

Structs

Conv2dBackward

Gradient computed during the backward pass for each tensor used by conv2d.

Conv3dBackward

Gradient computed during the backward pass for each tensor used by conv3d.

ConvOptions

Convolution options.

ConvTransposeOptions

Transposed convolution options.

DeformConv2dBackward

Gradient computed during the backward pass for each tensor used by deform_conv2d.

DeformConvOptions

Convolution options.

InterpolateBackward

Gradient computed during the backward pass for each tensor used by interpolate.

InterpolateOptions

Interpolation options.

MaxPool1dBackward

Gradient computed during the backward pass for each tensor used by max_pool1d.

MaxPool1dWithIndices

Results from max_pool1d.

MaxPool2dBackward

Gradient computed during the backward pass for each tensor used by max_pool2d.

MaxPool2dWithIndices

Results from max_pool2d.

TransactionPrimitive

Contains all tensor primitives that are going to be read.

TransactionPrimitiveResult

Contains all data related to a transaction.

UnfoldOptions

Unfold operation options.

Enums

InterpolateMode

Algorithm used for upsampling.

Traits

ActivationOps

Activation function operations.

BoolTensorOps

Bool Tensor API for basic operations, see tensor for documentation on each function.

FloatTensorOps

Operations on float tensors.

IntTensorOps

Int Tensor API for basic and numeric operations, see tensor for documentation on each function.

ModuleOps

Module operations trait.

QTensorOps

Quantized Tensor API for basic operations, see tensor for documentation on each function.

TransactionOps

Operations that are sync by nature and that can be batch together in transactions to improve compute utilization with efficient laziness.

Functions

binary_ops_shape

Computes the output shape for binary operations with broadcasting support.

Type Aliases

BoolTensor

Boolean tensor primitive type used by the backend.

Device

Device type used by the backend.

FloatElem

Float element type used by backend.

FloatTensor

Float tensor primitive type used by the backend.

IntElem

Integer element type used by backend.

IntTensor

Integer tensor primitive type used by the backend.

QuantizedTensor

Quantized tensor primitive type used by the backend.