Trait burn_core::tensor::ops::BoolTensorOps

pub trait BoolTensorOps<B>where
    B: Backend,{

    // Required methods
    fn bool_empty<const D: usize>(
        shape: Shape<D>,
        device: &<B as Backend>::Device
    ) -> <B as Backend>::BoolTensorPrimitive<D>;
    fn bool_shape<const D: usize>(
        tensor: &<B as Backend>::BoolTensorPrimitive<D>
    ) -> Shape<D>;
    fn bool_into_data<const D: usize>(
        tensor: <B as Backend>::BoolTensorPrimitive<D>
    ) -> Reader<Data<bool, D>>;
    fn bool_from_data<const D: usize>(
        data: Data<bool, D>,
        device: &<B as Backend>::Device
    ) -> <B as Backend>::BoolTensorPrimitive<D>;
    fn bool_into_int<const D: usize>(
        tensor: <B as Backend>::BoolTensorPrimitive<D>
    ) -> <B as Backend>::IntTensorPrimitive<D>;
    fn bool_into_float<const D: usize>(
        tensor: <B as Backend>::BoolTensorPrimitive<D>
    ) -> <B as Backend>::TensorPrimitive<D>;
    fn bool_device<const D: usize>(
        tensor: &<B as Backend>::BoolTensorPrimitive<D>
    ) -> <B as Backend>::Device;
    fn bool_to_device<const D: usize>(
        tensor: <B as Backend>::BoolTensorPrimitive<D>,
        device: &<B as Backend>::Device
    ) -> <B as Backend>::BoolTensorPrimitive<D>;
    fn bool_reshape<const D1: usize, const D2: usize>(
        tensor: <B as Backend>::BoolTensorPrimitive<D1>,
        shape: Shape<D2>
    ) -> <B as Backend>::BoolTensorPrimitive<D2>;
    fn bool_slice<const D1: usize, const D2: usize>(
        tensor: <B as Backend>::BoolTensorPrimitive<D1>,
        ranges: [Range<usize>; D2]
    ) -> <B as Backend>::BoolTensorPrimitive<D1>;
    fn bool_slice_assign<const D1: usize, const D2: usize>(
        tensor: <B as Backend>::BoolTensorPrimitive<D1>,
        ranges: [Range<usize>; D2],
        value: <B as Backend>::BoolTensorPrimitive<D1>
    ) -> <B as Backend>::BoolTensorPrimitive<D1>;
    fn bool_cat<const D: usize>(
        tensors: Vec<<B as Backend>::BoolTensorPrimitive<D>, Global>,
        dim: usize
    ) -> <B as Backend>::BoolTensorPrimitive<D>;
    fn bool_equal<const D: usize>(
        lhs: <B as Backend>::BoolTensorPrimitive<D>,
        rhs: <B as Backend>::BoolTensorPrimitive<D>
    ) -> <B as Backend>::BoolTensorPrimitive<D>;
    fn bool_not<const D: usize>(
        tensor: <B as Backend>::BoolTensorPrimitive<D>
    ) -> <B as Backend>::BoolTensorPrimitive<D>;
    fn bool_swap_dims<const D: usize>(
        tensor: <B as Backend>::BoolTensorPrimitive<D>,
        dim1: usize,
        dim2: usize
    ) -> <B as Backend>::BoolTensorPrimitive<D>;

    // Provided methods
    fn bool_to_data<const D: usize>(
        tensor: &<B as Backend>::BoolTensorPrimitive<D>
    ) -> Reader<Data<bool, D>> { ... }
    fn bool_repeat<const D: usize>(
        tensor: <B as Backend>::BoolTensorPrimitive<D>,
        dim: usize,
        times: usize
    ) -> <B as Backend>::BoolTensorPrimitive<D> { ... }
    fn bool_transpose<const D: usize>(
        tensor: <B as Backend>::BoolTensorPrimitive<D>
    ) -> <B as Backend>::BoolTensorPrimitive<D> { ... }

}

Bool Tensor API for basic operations, see tensor for documentation on each function.

Required Methods

fn bool_empty<const D: usize>( shape: Shape<D>, device: &<B as Backend>::Device ) -> <B as Backend>::BoolTensorPrimitive<D>

Creates a new bool tensor.

Arguments

shape - The shape of the tensor.

device - The device to create the tensor on.

Returns

The boolean tensor with the given shape.

fn bool_shape<const D: usize>( tensor: &<B as Backend>::BoolTensorPrimitive<D> ) -> Shape<D>

Returns the shape of the tensor.

Arguments

tensor - The tensor.

Returns

The shape of the tensor.

fn bool_into_data<const D: usize>( tensor: <B as Backend>::BoolTensorPrimitive<D> ) -> Reader<Data<bool, D>>

Converts the tensor to a data structure.

Arguments

tensor - The tensor.

Returns

The data structure with the tensor’s data.

fn bool_from_data<const D: usize>( data: Data<bool, D>, device: &<B as Backend>::Device ) -> <B as Backend>::BoolTensorPrimitive<D>

Creates a tensor from the data structure.

Arguments

data - The data structure.

device - The device to create the tensor on.

Returns

The tensor with the data.

fn bool_into_int<const D: usize>( tensor: <B as Backend>::BoolTensorPrimitive<D> ) -> <B as Backend>::IntTensorPrimitive<D>

Converts bool tensor to int tensor.

Arguments

tensor - The tensor.

Returns

The int tensor with the same data as the bool tensor.

fn bool_into_float<const D: usize>( tensor: <B as Backend>::BoolTensorPrimitive<D> ) -> <B as Backend>::TensorPrimitive<D>

Converts bool tensor to float tensor.

Arguments

tensor - The tensor.

Returns

The float tensor with the same data as the bool tensor.

fn bool_device<const D: usize>( tensor: &<B as Backend>::BoolTensorPrimitive<D> ) -> <B as Backend>::Device

Gets the device of the tensor.

Arguments

tensor - The tensor.

Returns

The device of the tensor.

fn bool_to_device<const D: usize>( tensor: <B as Backend>::BoolTensorPrimitive<D>, device: &<B as Backend>::Device ) -> <B as Backend>::BoolTensorPrimitive<D>

Moves the tensor to the device.

fn bool_reshape<const D1: usize, const D2: usize>( tensor: <B as Backend>::BoolTensorPrimitive<D1>, shape: Shape<D2> ) -> <B as Backend>::BoolTensorPrimitive<D2>

Reshapes the tensor.

Arguments

tensor - The tensor.

shape - The new shape.

Returns

The tensor with the new shape.

fn bool_slice<const D1: usize, const D2: usize>( tensor: <B as Backend>::BoolTensorPrimitive<D1>, ranges: [Range<usize>; D2] ) -> <B as Backend>::BoolTensorPrimitive<D1>

Gets the values from the tensor for the given ranges.

Arguments

tensor - The tensor.

ranges - The ranges to get the values from.

Returns

The tensor with the values for the given ranges.

fn bool_slice_assign<const D1: usize, const D2: usize>( tensor: <B as Backend>::BoolTensorPrimitive<D1>, ranges: [Range<usize>; D2], value: <B as Backend>::BoolTensorPrimitive<D1> ) -> <B as Backend>::BoolTensorPrimitive<D1>

Sets the values in the tensor for the given ranges.

Arguments

tensor - The tensor.

ranges - The ranges to set the values for.

value - The values to set.

Returns

The tensor with the values set for the given ranges.

fn bool_cat<const D: usize>( tensors: Vec<<B as Backend>::BoolTensorPrimitive<D>, Global>, dim: usize ) -> <B as Backend>::BoolTensorPrimitive<D>

Concatenates the tensors along the given dimension.

Arguments

tensors - The tensors to concatenate.

dim - The dimension to concatenate along.

Returns

The tensor with the tensors concatenated along the given dimension.

fn bool_equal<const D: usize>( lhs: <B as Backend>::BoolTensorPrimitive<D>, rhs: <B as Backend>::BoolTensorPrimitive<D> ) -> <B as Backend>::BoolTensorPrimitive<D>

Equates the two tensors.

Arguments

lhs - The left hand side tensor.

rhs - The right hand side tensor.

Returns

The tensor with the result of the equate.

fn bool_not<const D: usize>( tensor: <B as Backend>::BoolTensorPrimitive<D> ) -> <B as Backend>::BoolTensorPrimitive<D>

Inverses boolean values.

Arguments

tensor - The tensor.

Returns

The tensor with the result of the negation.

fn bool_swap_dims<const D: usize>( tensor: <B as Backend>::BoolTensorPrimitive<D>, dim1: usize, dim2: usize ) -> <B as Backend>::BoolTensorPrimitive<D>

Swaps two dimensions of a bool tensor.

Arguments

tensor - The tensor to swap the dimensions of.

dim1 - The first dimension to swap.

dim2 - The second dimension to swap.

Returns

The tensor with the dimensions swapped.

Provided Methods

fn bool_to_data<const D: usize>( tensor: &<B as Backend>::BoolTensorPrimitive<D> ) -> Reader<Data<bool, D>>

Gets the data from the tensor.

Arguments

data - The data structure.

Returns

The data cloned from the data structure.

fn bool_repeat<const D: usize>( tensor: <B as Backend>::BoolTensorPrimitive<D>, dim: usize, times: usize ) -> <B as Backend>::BoolTensorPrimitive<D>

Repeats one dimension of the tensor a given number of times along that dimension.

Arguments

tensor - The tensor.

dim - The dimension to repeat.

times - The number of times to repeat the dimension.

Returns

The tensor with the dimension repeated.

fn bool_transpose<const D: usize>( tensor: <B as Backend>::BoolTensorPrimitive<D> ) -> <B as Backend>::BoolTensorPrimitive<D>

Transposes a bool tensor.

Arguments

tensor - The tensor to transpose.

Returns

The transposed tensor.