Trait burn_core::tensor::ops::TensorOps

pub trait TensorOps<B>where
    B: Backend,{

    // Required methods
    fn from_data<const D: usize>(
        data: Data<<B as Backend>::FloatElem, D>,
        device: &<B as Backend>::Device
    ) -> <B as Backend>::TensorPrimitive<D>;
    fn random<const D: usize>(
        shape: Shape<D>,
        distribution: Distribution<<B as Backend>::FloatElem>,
        device: &<B as Backend>::Device
    ) -> <B as Backend>::TensorPrimitive<D>;
    fn shape<const D: usize>(
        tensor: &<B as Backend>::TensorPrimitive<D>
    ) -> Shape<D>;
    fn into_data<const D: usize>(
        tensor: <B as Backend>::TensorPrimitive<D>
    ) -> Reader<Data<<B as Backend>::FloatElem, D>>;
    fn device<const D: usize>(
        tensor: &<B as Backend>::TensorPrimitive<D>
    ) -> <B as Backend>::Device;
    fn to_device<const D: usize>(
        tensor: <B as Backend>::TensorPrimitive<D>,
        device: &<B as Backend>::Device
    ) -> <B as Backend>::TensorPrimitive<D>;
    fn into_int<const D: usize>(
        tensor: <B as Backend>::TensorPrimitive<D>
    ) -> <B as Backend>::IntTensorPrimitive<D>;
    fn empty<const D: usize>(
        shape: Shape<D>,
        device: &<B as Backend>::Device
    ) -> <B as Backend>::TensorPrimitive<D>;
    fn add<const D: usize>(
        lhs: <B as Backend>::TensorPrimitive<D>,
        rhs: <B as Backend>::TensorPrimitive<D>
    ) -> <B as Backend>::TensorPrimitive<D>;
    fn add_scalar<const D: usize>(
        lhs: <B as Backend>::TensorPrimitive<D>,
        rhs: <B as Backend>::FloatElem
    ) -> <B as Backend>::TensorPrimitive<D>;
    fn sub<const D: usize>(
        lhs: <B as Backend>::TensorPrimitive<D>,
        rhs: <B as Backend>::TensorPrimitive<D>
    ) -> <B as Backend>::TensorPrimitive<D>;
    fn sub_scalar<const D: usize>(
        lhs: <B as Backend>::TensorPrimitive<D>,
        rhs: <B as Backend>::FloatElem
    ) -> <B as Backend>::TensorPrimitive<D>;
    fn mul<const D: usize>(
        lhs: <B as Backend>::TensorPrimitive<D>,
        rhs: <B as Backend>::TensorPrimitive<D>
    ) -> <B as Backend>::TensorPrimitive<D>;
    fn mul_scalar<const D: usize>(
        lhs: <B as Backend>::TensorPrimitive<D>,
        rhs: <B as Backend>::FloatElem
    ) -> <B as Backend>::TensorPrimitive<D>;
    fn div<const D: usize>(
        lhs: <B as Backend>::TensorPrimitive<D>,
        rhs: <B as Backend>::TensorPrimitive<D>
    ) -> <B as Backend>::TensorPrimitive<D>;
    fn div_scalar<const D: usize>(
        lhs: <B as Backend>::TensorPrimitive<D>,
        rhs: <B as Backend>::FloatElem
    ) -> <B as Backend>::TensorPrimitive<D>;
    fn matmul<const D: usize>(
        lhs: <B as Backend>::TensorPrimitive<D>,
        rhs: <B as Backend>::TensorPrimitive<D>
    ) -> <B as Backend>::TensorPrimitive<D>;
    fn swap_dims<const D: usize>(
        tensor: <B as Backend>::TensorPrimitive<D>,
        dim1: usize,
        dim2: usize
    ) -> <B as Backend>::TensorPrimitive<D>;
    fn reshape<const D1: usize, const D2: usize>(
        tensor: <B as Backend>::TensorPrimitive<D1>,
        shape: Shape<D2>
    ) -> <B as Backend>::TensorPrimitive<D2>;
    fn gather<const D: usize>(
        dim: usize,
        tensor: <B as Backend>::TensorPrimitive<D>,
        indices: <B as Backend>::IntTensorPrimitive<D>
    ) -> <B as Backend>::TensorPrimitive<D>;
    fn scatter<const D: usize>(
        dim: usize,
        tensor: <B as Backend>::TensorPrimitive<D>,
        indices: <B as Backend>::IntTensorPrimitive<D>,
        value: <B as Backend>::TensorPrimitive<D>
    ) -> <B as Backend>::TensorPrimitive<D>;
    fn select<const D: usize>(
        tensor: <B as Backend>::TensorPrimitive<D>,
        dim: usize,
        indices: <B as Backend>::IntTensorPrimitive<1>
    ) -> <B as Backend>::TensorPrimitive<D>;
    fn select_assign<const D: usize>(
        tensor: <B as Backend>::TensorPrimitive<D>,
        dim: usize,
        indices: <B as Backend>::IntTensorPrimitive<1>,
        value: <B as Backend>::TensorPrimitive<D>
    ) -> <B as Backend>::TensorPrimitive<D>;
    fn slice<const D1: usize, const D2: usize>(
        tensor: <B as Backend>::TensorPrimitive<D1>,
        ranges: [Range<usize>; D2]
    ) -> <B as Backend>::TensorPrimitive<D1>;
    fn slice_assign<const D1: usize, const D2: usize>(
        tensor: <B as Backend>::TensorPrimitive<D1>,
        ranges: [Range<usize>; D2],
        value: <B as Backend>::TensorPrimitive<D1>
    ) -> <B as Backend>::TensorPrimitive<D1>;
    fn mask_where<const D: usize>(
        tensor: <B as Backend>::TensorPrimitive<D>,
        mask: <B as Backend>::BoolTensorPrimitive<D>,
        value: <B as Backend>::TensorPrimitive<D>
    ) -> <B as Backend>::TensorPrimitive<D>;
    fn mask_fill<const D: usize>(
        tensor: <B as Backend>::TensorPrimitive<D>,
        mask: <B as Backend>::BoolTensorPrimitive<D>,
        value: <B as Backend>::FloatElem
    ) -> <B as Backend>::TensorPrimitive<D>;
    fn equal<const D: usize>(
        lhs: <B as Backend>::TensorPrimitive<D>,
        rhs: <B as Backend>::TensorPrimitive<D>
    ) -> <B as Backend>::BoolTensorPrimitive<D>;
    fn equal_elem<const D: usize>(
        lhs: <B as Backend>::TensorPrimitive<D>,
        rhs: <B as Backend>::FloatElem
    ) -> <B as Backend>::BoolTensorPrimitive<D>;
    fn greater<const D: usize>(
        lhs: <B as Backend>::TensorPrimitive<D>,
        rhs: <B as Backend>::TensorPrimitive<D>
    ) -> <B as Backend>::BoolTensorPrimitive<D>;
    fn greater_elem<const D: usize>(
        lhs: <B as Backend>::TensorPrimitive<D>,
        rhs: <B as Backend>::FloatElem
    ) -> <B as Backend>::BoolTensorPrimitive<D>;
    fn greater_equal<const D: usize>(
        lhs: <B as Backend>::TensorPrimitive<D>,
        rhs: <B as Backend>::TensorPrimitive<D>
    ) -> <B as Backend>::BoolTensorPrimitive<D>;
    fn greater_equal_elem<const D: usize>(
        lhs: <B as Backend>::TensorPrimitive<D>,
        rhs: <B as Backend>::FloatElem
    ) -> <B as Backend>::BoolTensorPrimitive<D>;
    fn lower<const D: usize>(
        lhs: <B as Backend>::TensorPrimitive<D>,
        rhs: <B as Backend>::TensorPrimitive<D>
    ) -> <B as Backend>::BoolTensorPrimitive<D>;
    fn lower_elem<const D: usize>(
        lhs: <B as Backend>::TensorPrimitive<D>,
        rhs: <B as Backend>::FloatElem
    ) -> <B as Backend>::BoolTensorPrimitive<D>;
    fn lower_equal<const D: usize>(
        lhs: <B as Backend>::TensorPrimitive<D>,
        rhs: <B as Backend>::TensorPrimitive<D>
    ) -> <B as Backend>::BoolTensorPrimitive<D>;
    fn lower_equal_elem<const D: usize>(
        lhs: <B as Backend>::TensorPrimitive<D>,
        rhs: <B as Backend>::FloatElem
    ) -> <B as Backend>::BoolTensorPrimitive<D>;
    fn sum<const D: usize>(
        tensor: <B as Backend>::TensorPrimitive<D>
    ) -> <B as Backend>::TensorPrimitive<1>;
    fn sum_dim<const D: usize>(
        tensor: <B as Backend>::TensorPrimitive<D>,
        dim: usize
    ) -> <B as Backend>::TensorPrimitive<D>;
    fn mean_dim<const D: usize>(
        tensor: <B as Backend>::TensorPrimitive<D>,
        dim: usize
    ) -> <B as Backend>::TensorPrimitive<D>;
    fn to_full_precision<const D: usize>(
        tensor: &<B as Backend>::TensorPrimitive<D>
    ) -> <<B as Backend>::FullPrecisionBackend as Backend>::TensorPrimitive<D>;
    fn from_full_precision<const D: usize>(
        tensor: <<B as Backend>::FullPrecisionBackend as Backend>::TensorPrimitive<D>
    ) -> <B as Backend>::TensorPrimitive<D>;
    fn exp<const D: usize>(
        tensor: <B as Backend>::TensorPrimitive<D>
    ) -> <B as Backend>::TensorPrimitive<D>;
    fn log<const D: usize>(
        tensor: <B as Backend>::TensorPrimitive<D>
    ) -> <B as Backend>::TensorPrimitive<D>;
    fn log1p<const D: usize>(
        tensor: <B as Backend>::TensorPrimitive<D>
    ) -> <B as Backend>::TensorPrimitive<D>;
    fn powf<const D: usize>(
        tensor: <B as Backend>::TensorPrimitive<D>,
        value: f32
    ) -> <B as Backend>::TensorPrimitive<D>;
    fn sqrt<const D: usize>(
        tensor: <B as Backend>::TensorPrimitive<D>
    ) -> <B as Backend>::TensorPrimitive<D>;
    fn abs<const D: usize>(
        tensor: <B as Backend>::TensorPrimitive<D>
    ) -> <B as Backend>::TensorPrimitive<D>;
    fn cos<const D: usize>(
        tensor: <B as Backend>::TensorPrimitive<D>
    ) -> <B as Backend>::TensorPrimitive<D>;
    fn sin<const D: usize>(
        tensor: <B as Backend>::TensorPrimitive<D>
    ) -> <B as Backend>::TensorPrimitive<D>;
    fn tanh<const D: usize>(
        tensor: <B as Backend>::TensorPrimitive<D>
    ) -> <B as Backend>::TensorPrimitive<D>;
    fn erf<const D: usize>(
        tensor: <B as Backend>::TensorPrimitive<D>
    ) -> <B as Backend>::TensorPrimitive<D>;
    fn cat<const D: usize>(
        tensors: Vec<<B as Backend>::TensorPrimitive<D>, Global>,
        dim: usize
    ) -> <B as Backend>::TensorPrimitive<D>;
    fn argmax<const D: usize>(
        tensor: <B as Backend>::TensorPrimitive<D>,
        dim: usize
    ) -> <B as Backend>::IntTensorPrimitive<D>;
    fn argmin<const D: usize>(
        tensor: <B as Backend>::TensorPrimitive<D>,
        dim: usize
    ) -> <B as Backend>::IntTensorPrimitive<D>;

    // Provided methods
    fn zeros<const D: usize>(
        shape: Shape<D>,
        device: &<B as Backend>::Device
    ) -> <B as Backend>::TensorPrimitive<D> { ... }
    fn ones<const D: usize>(
        shape: Shape<D>,
        device: &<B as Backend>::Device
    ) -> <B as Backend>::TensorPrimitive<D> { ... }
    fn full<const D: usize>(
        shape: Shape<D>,
        fill_value: <B as Backend>::FloatElem,
        device: &<B as Backend>::Device
    ) -> <B as Backend>::TensorPrimitive<D> { ... }
    fn to_data<const D: usize>(
        tensor: &<B as Backend>::TensorPrimitive<D>
    ) -> Reader<Data<<B as Backend>::FloatElem, D>> { ... }
    fn arange(
        range: Range<usize>,
        device: &<B as Backend>::Device
    ) -> <B as Backend>::IntTensorPrimitive<1> { ... }
    fn arange_step(
        range: Range<usize>,
        step: usize,
        device: &<B as Backend>::Device
    ) -> <B as Backend>::IntTensorPrimitive<1> { ... }
    fn repeat<const D: usize>(
        tensor: <B as Backend>::TensorPrimitive<D>,
        dim: usize,
        times: usize
    ) -> <B as Backend>::TensorPrimitive<D> { ... }
    fn clamp_min<const D: usize>(
        tensor: <B as Backend>::TensorPrimitive<D>,
        min: <B as Backend>::FloatElem
    ) -> <B as Backend>::TensorPrimitive<D> { ... }
    fn clamp_max<const D: usize>(
        tensor: <B as Backend>::TensorPrimitive<D>,
        max: <B as Backend>::FloatElem
    ) -> <B as Backend>::TensorPrimitive<D> { ... }
    fn clamp<const D: usize>(
        tensor: <B as Backend>::TensorPrimitive<D>,
        min: <B as Backend>::FloatElem,
        max: <B as Backend>::FloatElem
    ) -> <B as Backend>::TensorPrimitive<D> { ... }
    fn neg<const D: usize>(
        tensor: <B as Backend>::TensorPrimitive<D>
    ) -> <B as Backend>::TensorPrimitive<D> { ... }
    fn transpose<const D: usize>(
        tensor: <B as Backend>::TensorPrimitive<D>
    ) -> <B as Backend>::TensorPrimitive<D> { ... }
    fn detach<const D: usize>(
        tensor: <B as Backend>::TensorPrimitive<D>
    ) -> <B as Backend>::TensorPrimitive<D> { ... }
    fn set_require_grad<const D: usize>(
        tensor: <B as Backend>::TensorPrimitive<D>,
        _require_grad: bool
    ) -> <B as Backend>::TensorPrimitive<D> { ... }
    fn is_require_grad<const D: usize>(
        _tensor: &<B as Backend>::TensorPrimitive<D>
    ) -> bool { ... }
    fn mean<const D: usize>(
        tensor: <B as Backend>::TensorPrimitive<D>
    ) -> <B as Backend>::TensorPrimitive<1> { ... }
    fn max<const D: usize>(
        tensor: <B as Backend>::TensorPrimitive<D>
    ) -> <B as Backend>::TensorPrimitive<1> { ... }
    fn max_dim<const D: usize>(
        tensor: <B as Backend>::TensorPrimitive<D>,
        dim: usize
    ) -> <B as Backend>::TensorPrimitive<D> { ... }
    fn max_dim_with_indices<const D: usize>(
        tensor: <B as Backend>::TensorPrimitive<D>,
        dim: usize
    ) -> (<B as Backend>::TensorPrimitive<D>, <B as Backend>::IntTensorPrimitive<D>) { ... }
    fn min<const D: usize>(
        tensor: <B as Backend>::TensorPrimitive<D>
    ) -> <B as Backend>::TensorPrimitive<1> { ... }
    fn min_dim<const D: usize>(
        tensor: <B as Backend>::TensorPrimitive<D>,
        dim: usize
    ) -> <B as Backend>::TensorPrimitive<D> { ... }
    fn min_dim_with_indices<const D: usize>(
        tensor: <B as Backend>::TensorPrimitive<D>,
        dim: usize
    ) -> (<B as Backend>::TensorPrimitive<D>, <B as Backend>::IntTensorPrimitive<D>) { ... }

}

Operations on float tensors.

Required Methods

fn from_data<const D: usize>( data: Data<<B as Backend>::FloatElem, D>, device: &<B as Backend>::Device ) -> <B as Backend>::TensorPrimitive<D>

Creates a new tensor from the data structure.

Arguments

data - The data structure.

device - The device to create the tensor on.

Returns

The tensor with the given data.

fn random<const D: usize>( shape: Shape<D>, distribution: Distribution<<B as Backend>::FloatElem>, device: &<B as Backend>::Device ) -> <B as Backend>::TensorPrimitive<D>

Creates a new tensor with random values.

Arguments

shape - The shape of the tensor.

distribution - The distribution to sample from.

device - The device to create the tensor on.

Returns

The tensor with the given shape and random values.

fn shape<const D: usize>( tensor: &<B as Backend>::TensorPrimitive<D> ) -> Shape<D>

Gets the shape of the tensor.

Arguments

tensor - The tensor.

Returns

The shape of the tensor.

fn into_data<const D: usize>( tensor: <B as Backend>::TensorPrimitive<D> ) -> Reader<Data<<B as Backend>::FloatElem, D>>

Converts the tensor to a data structure.

Arguments

tensor - The tensor.

Returns

The data structure with the tensor’s data.

fn device<const D: usize>( tensor: &<B as Backend>::TensorPrimitive<D> ) -> <B as Backend>::Device

Gets the device of the tensor.

Arguments

tensor - The tensor.

Returns

The device of the tensor.

fn to_device<const D: usize>( tensor: <B as Backend>::TensorPrimitive<D>, device: &<B as Backend>::Device ) -> <B as Backend>::TensorPrimitive<D>

Moves the tensor to the given device.

Arguments

tensor - The tensor.

device - The device to move the tensor to.

Returns

The tensor on the given device.

fn into_int<const D: usize>( tensor: <B as Backend>::TensorPrimitive<D> ) -> <B as Backend>::IntTensorPrimitive<D>

Converts float tensor to int tensor.

Arguments

tensor - The tensor.

Returns

The int tensor with the same data as the float tensor.

fn empty<const D: usize>( shape: Shape<D>, device: &<B as Backend>::Device ) -> <B as Backend>::TensorPrimitive<D>

Creates an empty tensor with the given shape.

Arguments

shape - The shape of the tensor.

device - The device to create the tensor on.

Returns

The empty tensor with the given shape.

fn add<const D: usize>( lhs: <B as Backend>::TensorPrimitive<D>, rhs: <B as Backend>::TensorPrimitive<D> ) -> <B as Backend>::TensorPrimitive<D>

Adds two tensors together.

Arguments

lhs - The left hand side tensor.

rhs - The right hand side tensor.

Returns

The result of adding the two tensors together.

fn add_scalar<const D: usize>( lhs: <B as Backend>::TensorPrimitive<D>, rhs: <B as Backend>::FloatElem ) -> <B as Backend>::TensorPrimitive<D>

Adds a scalar to a tensor.

Arguments

lhs - The left hand side tensor.

rhs - The right hand side scalar.

Returns

The result of adding the scalar to the tensor.

fn sub<const D: usize>( lhs: <B as Backend>::TensorPrimitive<D>, rhs: <B as Backend>::TensorPrimitive<D> ) -> <B as Backend>::TensorPrimitive<D>

Subtracts two tensors.

Arguments

lhs - The left hand side tensor.

rhs - The right hand side tensor.

Returns

The result of subtracting the two tensors.

fn sub_scalar<const D: usize>( lhs: <B as Backend>::TensorPrimitive<D>, rhs: <B as Backend>::FloatElem ) -> <B as Backend>::TensorPrimitive<D>

Subtracts a scalar from a tensor.

Arguments

lhs - The left hand side tensor.

rhs - The right hand side scalar.

Returns

The result of subtracting the scalar from the tensor.

fn mul<const D: usize>( lhs: <B as Backend>::TensorPrimitive<D>, rhs: <B as Backend>::TensorPrimitive<D> ) -> <B as Backend>::TensorPrimitive<D>

Multiplies two tensors together element-wise.

fn mul_scalar<const D: usize>( lhs: <B as Backend>::TensorPrimitive<D>, rhs: <B as Backend>::FloatElem ) -> <B as Backend>::TensorPrimitive<D>

Multiplies a tensor by a scalar.

Arguments

lhs - The left hand side tensor.

rhs - The right hand side scalar.

Returns

The result of multiplying the tensor by the scalar.

fn div<const D: usize>( lhs: <B as Backend>::TensorPrimitive<D>, rhs: <B as Backend>::TensorPrimitive<D> ) -> <B as Backend>::TensorPrimitive<D>

Divides two tensors element-wise.

Arguments

lhs - The left hand side tensor.

rhs - The right hand side tensor.

Returns

The result of dividing the two tensors.

fn div_scalar<const D: usize>( lhs: <B as Backend>::TensorPrimitive<D>, rhs: <B as Backend>::FloatElem ) -> <B as Backend>::TensorPrimitive<D>

Divides a tensor by a scalar.

Arguments

lhs - The left hand side tensor.

rhs - The right hand side scalar.

Returns

The result of dividing the tensor by the scalar.

fn matmul<const D: usize>( lhs: <B as Backend>::TensorPrimitive<D>, rhs: <B as Backend>::TensorPrimitive<D> ) -> <B as Backend>::TensorPrimitive<D>

Multiplies two tensors together using matrix multiplication.

Arguments

lhs - The left hand side tensor.

rhs - The right hand side tensor.

Returns

The result of multiplying the two tensors together using matrix multiplication.

fn swap_dims<const D: usize>( tensor: <B as Backend>::TensorPrimitive<D>, dim1: usize, dim2: usize ) -> <B as Backend>::TensorPrimitive<D>

Swaps two dimensions of a tensor.

Arguments

tensor - The tensor to swap the dimensions of.

dim1 - The first dimension to swap.

dim2 - The second dimension to swap.

Returns

The tensor with the dimensions swapped.

fn reshape<const D1: usize, const D2: usize>( tensor: <B as Backend>::TensorPrimitive<D1>, shape: Shape<D2> ) -> <B as Backend>::TensorPrimitive<D2>

Reshapes a tensor.

Arguments

tensor - The tensor to reshape.

shape - The new shape of the tensor.

Returns

The tensor with the new shape.

fn gather<const D: usize>( dim: usize, tensor: <B as Backend>::TensorPrimitive<D>, indices: <B as Backend>::IntTensorPrimitive<D> ) -> <B as Backend>::TensorPrimitive<D>

Gather elements from a tensor.

Arguments

dim - The dimension to gather from.

tensor - The tensor to gather from.

indices - The indices to gather.

Returns

The gathered elements.

fn scatter<const D: usize>( dim: usize, tensor: <B as Backend>::TensorPrimitive<D>, indices: <B as Backend>::IntTensorPrimitive<D>, value: <B as Backend>::TensorPrimitive<D> ) -> <B as Backend>::TensorPrimitive<D>

Scatter elements into a tensor.

Arguments

dim - The dimension to scatter into.

tensor - The tensor to scatter into.

indices - The indices to scatter into.

value - The value to scatter.

Returns

The tensor with the scattered elements.

fn select<const D: usize>( tensor: <B as Backend>::TensorPrimitive<D>, dim: usize, indices: <B as Backend>::IntTensorPrimitive<1> ) -> <B as Backend>::TensorPrimitive<D>

Select tensor elements along the given dimension corresponding for the given indices.

Arguments

tensor - The tensor to select from.

dim - The dimension to select from.

indices - The indices to select.

Returns

The selected elements.

fn select_assign<const D: usize>( tensor: <B as Backend>::TensorPrimitive<D>, dim: usize, indices: <B as Backend>::IntTensorPrimitive<1>, value: <B as Backend>::TensorPrimitive<D> ) -> <B as Backend>::TensorPrimitive<D>

Assign the selected elements along the given dimension corresponding for the given indices to the given value.

Arguments

tensor - The tensor to select from.

dim - The dimension to select from.

indices - The indices to select.

value - The value to assign.

Returns

The tensor with the selected elements assigned to the given value.

fn slice<const D1: usize, const D2: usize>( tensor: <B as Backend>::TensorPrimitive<D1>, ranges: [Range<usize>; D2] ) -> <B as Backend>::TensorPrimitive<D1>

Select tensor elements corresponding for the given ranges.

Arguments

tensor - The tensor to select from.

ranges - The ranges to select.

Returns

The selected elements in a new tensor.

fn slice_assign<const D1: usize, const D2: usize>( tensor: <B as Backend>::TensorPrimitive<D1>, ranges: [Range<usize>; D2], value: <B as Backend>::TensorPrimitive<D1> ) -> <B as Backend>::TensorPrimitive<D1>

Assign the selected elements corresponding for the given ranges to the given value.

Arguments

tensor - The tensor to select from.

ranges - The ranges to select.

value - The value to assign.

Returns

The tensor with the selected elements assigned to the given value.

fn mask_where<const D: usize>( tensor: <B as Backend>::TensorPrimitive<D>, mask: <B as Backend>::BoolTensorPrimitive<D>, value: <B as Backend>::TensorPrimitive<D> ) -> <B as Backend>::TensorPrimitive<D>

Update the given tensor with the value tensor where the mask is true.

Arguments

tensor - The tensor to select from.

mask - The boolean mask to select with.

value - The value to assign to the selected elements from the value tensor.

Returns

The tensor with the selected elements assigned to the given value.

fn mask_fill<const D: usize>( tensor: <B as Backend>::TensorPrimitive<D>, mask: <B as Backend>::BoolTensorPrimitive<D>, value: <B as Backend>::FloatElem ) -> <B as Backend>::TensorPrimitive<D>

Update the given tensor with the value where the mask is true.

Arguments

tensor - The tensor to select from.

mask - The boolean mask to select with.

value - The value to assign to the selected elements.

Returns

The tensor with the selected elements assigned to the given value.

fn equal<const D: usize>( lhs: <B as Backend>::TensorPrimitive<D>, rhs: <B as Backend>::TensorPrimitive<D> ) -> <B as Backend>::BoolTensorPrimitive<D>

Equal comparison of two tensors.

Arguments

lhs - The left hand side tensor.

rhs - The right hand side tensor.

Returns

A boolean tensor with the result of the comparison.

fn equal_elem<const D: usize>( lhs: <B as Backend>::TensorPrimitive<D>, rhs: <B as Backend>::FloatElem ) -> <B as Backend>::BoolTensorPrimitive<D>

Equal comparison of a tensor and a scalar.

Arguments

lhs - The left hand side tensor.

rhs - The right hand side scalar.

Returns

A boolean tensor with the result of the comparison.

fn greater<const D: usize>( lhs: <B as Backend>::TensorPrimitive<D>, rhs: <B as Backend>::TensorPrimitive<D> ) -> <B as Backend>::BoolTensorPrimitive<D>

Greater than comparison of two tensors.

Arguments

lhs - The left hand side tensor.

rhs - The right hand side tensor.

Returns

A boolean tensor with the result of the comparison.

fn greater_elem<const D: usize>( lhs: <B as Backend>::TensorPrimitive<D>, rhs: <B as Backend>::FloatElem ) -> <B as Backend>::BoolTensorPrimitive<D>

Greater than comparison of a tensor and a scalar.

Arguments

lhs - The left hand side tensor.

rhs - The right hand side scalar.

Returns

A boolean tensor with the result of the comparison.

fn greater_equal<const D: usize>( lhs: <B as Backend>::TensorPrimitive<D>, rhs: <B as Backend>::TensorPrimitive<D> ) -> <B as Backend>::BoolTensorPrimitive<D>

Greater than or equal comparison of two tensors.

Arguments

lhs - The left hand side tensor.

rhs - The right hand side tensor.

Returns

A boolean tensor with the result of the comparison.

fn greater_equal_elem<const D: usize>( lhs: <B as Backend>::TensorPrimitive<D>, rhs: <B as Backend>::FloatElem ) -> <B as Backend>::BoolTensorPrimitive<D>

Greater than or equal comparison of a tensor and a scalar.

Arguments

lhs - The left hand side tensor.

rhs - The right hand side scalar.

Returns

A boolean tensor with the result of the comparison.

fn lower<const D: usize>( lhs: <B as Backend>::TensorPrimitive<D>, rhs: <B as Backend>::TensorPrimitive<D> ) -> <B as Backend>::BoolTensorPrimitive<D>

Less than comparison of two tensors.

Arguments

lhs - The left hand side tensor.

rhs - The right hand side tensor.

Returns

A boolean tensor with the result of the comparison.

fn lower_elem<const D: usize>( lhs: <B as Backend>::TensorPrimitive<D>, rhs: <B as Backend>::FloatElem ) -> <B as Backend>::BoolTensorPrimitive<D>

Less than comparison of a tensor and a scalar.

Arguments

lhs - The left hand side tensor.

rhs - The right hand side scalar.

Returns

A boolean tensor with the result of the comparison.

fn lower_equal<const D: usize>( lhs: <B as Backend>::TensorPrimitive<D>, rhs: <B as Backend>::TensorPrimitive<D> ) -> <B as Backend>::BoolTensorPrimitive<D>

Less than or equal comparison of two tensors.

Arguments

lhs - The left hand side tensor.

rhs - The right hand side tensor.

Returns

A boolean tensor with the result of the comparison.

fn lower_equal_elem<const D: usize>( lhs: <B as Backend>::TensorPrimitive<D>, rhs: <B as Backend>::FloatElem ) -> <B as Backend>::BoolTensorPrimitive<D>

Less than or equal comparison of a tensor and a scalar.

Arguments

lhs - The left hand side tensor.

rhs - The right hand side scalar.

Returns

A boolean tensor with the result of the comparison.

fn sum<const D: usize>( tensor: <B as Backend>::TensorPrimitive<D> ) -> <B as Backend>::TensorPrimitive<1>

Sum of all elements in a tensor.

Arguments

tensor - The tensor to sum.

Returns

A scalar tensor with the sum of all elements in tensor.

fn sum_dim<const D: usize>( tensor: <B as Backend>::TensorPrimitive<D>, dim: usize ) -> <B as Backend>::TensorPrimitive<D>

Sum of all elements in a tensor along a dimension.

Arguments

tensor - The tensor to sum.

dim - The dimension along which to sum.

Returns

A tensor with the sum of all elements in tensor along dim.

fn mean_dim<const D: usize>( tensor: <B as Backend>::TensorPrimitive<D>, dim: usize ) -> <B as Backend>::TensorPrimitive<D>

Mean of all elements in a tensor along a dimension.

Arguments

tensor - The tensor to mean.

dim - The dimension along which to mean.

Returns

A tensor with the mean of all elements in tensor along dim.

fn to_full_precision<const D: usize>( tensor: &<B as Backend>::TensorPrimitive<D> ) -> <<B as Backend>::FullPrecisionBackend as Backend>::TensorPrimitive<D>

Converts a tensor to full precision.

Arguments

tensor - The tensor to convert.

Returns

A tensor with the same values as tensor but with full precision.

fn from_full_precision<const D: usize>( tensor: <<B as Backend>::FullPrecisionBackend as Backend>::TensorPrimitive<D> ) -> <B as Backend>::TensorPrimitive<D>

Converts a tensor from full precision.

Arguments

tensor - The tensor to convert.

Returns

A tensor with the same values as tensor but with the precision of the backend.

fn exp<const D: usize>( tensor: <B as Backend>::TensorPrimitive<D> ) -> <B as Backend>::TensorPrimitive<D>

Returns a new tensor with exponential values.

Arguments

tensor - The tensor to exponentiate.

Returns

A tensor with the same shape as tensor with exponential values.

fn log<const D: usize>( tensor: <B as Backend>::TensorPrimitive<D> ) -> <B as Backend>::TensorPrimitive<D>

Returns a new tensor with natural logarithm values.

Arguments

tensor - The tensor to take the logarithm of.

Returns

A tensor with the same shape as tensor with natural logarithm values.

fn log1p<const D: usize>( tensor: <B as Backend>::TensorPrimitive<D> ) -> <B as Backend>::TensorPrimitive<D>

Returns a new tensor with logarithm values of (1 + Xi).

Arguments

tensor - The tensor to take the logarithm of.

Returns

A tensor with the same shape as tensor with logarithm values of (1 + Xi).

fn powf<const D: usize>( tensor: <B as Backend>::TensorPrimitive<D>, value: f32 ) -> <B as Backend>::TensorPrimitive<D>

Returns a new tensor with values raised to the power of value.

Arguments

tensor - The tensor to exponentiate.

value - The exponent.

Returns

A tensor with the same shape as tensor with values raised to the power of value.

fn sqrt<const D: usize>( tensor: <B as Backend>::TensorPrimitive<D> ) -> <B as Backend>::TensorPrimitive<D>

Returns a new tensor with square root values.

Arguments

tensor - The tensor to take the square root of.

Returns

A tensor with the same shape as tensor with square root values.

fn abs<const D: usize>( tensor: <B as Backend>::TensorPrimitive<D> ) -> <B as Backend>::TensorPrimitive<D>

Returns a new tensor with absolute values.

Arguments

tensor - The tensor to take absolute value of.

Returns

A tensor with the same shape as tensor with absolute values.

fn cos<const D: usize>( tensor: <B as Backend>::TensorPrimitive<D> ) -> <B as Backend>::TensorPrimitive<D>

Returns a new tensor with cosine values.

Arguments

tensor - The tensor to take the cosine of.

Returns

A tensor with the same shape as tensor with cosine values.

fn sin<const D: usize>( tensor: <B as Backend>::TensorPrimitive<D> ) -> <B as Backend>::TensorPrimitive<D>

Returns a new tensor with sine values.

Arguments

tensor - The tensor to take the sine of.

Returns

A tensor with the same shape as tensor with sine values.

fn tanh<const D: usize>( tensor: <B as Backend>::TensorPrimitive<D> ) -> <B as Backend>::TensorPrimitive<D>

Returns a new tensor with tangent values.

Arguments

tensor - The tensor to take the tangent of.

Returns

A tensor with the same shape as tensor with tangent values.

fn erf<const D: usize>( tensor: <B as Backend>::TensorPrimitive<D> ) -> <B as Backend>::TensorPrimitive<D>

Returns a new tensor with the error function values.

Arguments

tensor - The tensor to take the error function of.

Returns

A tensor with the same shape as tensor with error function values.

fn cat<const D: usize>( tensors: Vec<<B as Backend>::TensorPrimitive<D>, Global>, dim: usize ) -> <B as Backend>::TensorPrimitive<D>

Catcatenates tensors along a dimension.

Arguments

tensors - The tensors to catcatenate.

dim - The dimension along which to catcatenate.

Returns

A tensor with the catcatenated tensors along dim.

fn argmax<const D: usize>( tensor: <B as Backend>::TensorPrimitive<D>, dim: usize ) -> <B as Backend>::IntTensorPrimitive<D>

Gets the indices of the maximum elements of a tensor along an axis.

Arguments

tensor - The tensor to get the maximum elements of.

dim - The dimension along which to get the maximum elements.

Returns

A tensor with the indices of the maximum elements of tensor along dim.

fn argmin<const D: usize>( tensor: <B as Backend>::TensorPrimitive<D>, dim: usize ) -> <B as Backend>::IntTensorPrimitive<D>

Gets the indices of the minimum elements of a tensor along an axis.

Arguments

tensor - The tensor to get the minimum elements of.

dim - The dimension along which to get the minimum elements.

Returns

A tensor with the indices of the minimum elements of tensor along dim.

Provided Methods

fn zeros<const D: usize>( shape: Shape<D>, device: &<B as Backend>::Device ) -> <B as Backend>::TensorPrimitive<D>

Creates a new tensor with zeros.

Arguments

shape - The shape of the tensor.

device - The device to create the tensor on.

Returns

The tensor with the given shape and zeros.

fn ones<const D: usize>( shape: Shape<D>, device: &<B as Backend>::Device ) -> <B as Backend>::TensorPrimitive<D>

Creates a new tensor with ones.

Arguments

shape - The shape of the tensor.

device - The device to create the tensor on.

Returns

The tensor with the given shape and ones.

fn full<const D: usize>( shape: Shape<D>, fill_value: <B as Backend>::FloatElem, device: &<B as Backend>::Device ) -> <B as Backend>::TensorPrimitive<D>

Creates a tensor filled with given value.

Arguments

shape - The shape of the tensor.

fill_value - The value with which to fill the tensor.

device - The device to create the tensor on.

Returns

The tensor filled with given value

fn to_data<const D: usize>( tensor: &<B as Backend>::TensorPrimitive<D> ) -> Reader<Data<<B as Backend>::FloatElem, D>>

Converts the tensor to a data structure.

Arguments

tensor - The tensor.

Returns

The data structure with the tensor’s data.

fn arange( range: Range<usize>, device: &<B as Backend>::Device ) -> <B as Backend>::IntTensorPrimitive<1>

Creates a new tensor with values from the given range.

Arguments

range - The range of values.

device - The device to create the tensor on.

Returns

The tensor with the given values.

Remarks

Uses arange_step with a step size of 1 under the hood.

fn arange_step( range: Range<usize>, step: usize, device: &<B as Backend>::Device ) -> <B as Backend>::IntTensorPrimitive<1>

Creates a new tensor with values from the given range with the given step size.

Arguments

range - The range of values.

step - The step size.

device - The device to create the tensor on.

Returns

The tensor with the given values.

fn repeat<const D: usize>( tensor: <B as Backend>::TensorPrimitive<D>, dim: usize, times: usize ) -> <B as Backend>::TensorPrimitive<D>

Repeat the tensor along the given dimension.

Arguments

tensor - The tensor.

dim - The dimension to repeat.

times - The number of times to repeat the dimension.

Returns

The tensor with the given dimension repeated.

fn clamp_min<const D: usize>( tensor: <B as Backend>::TensorPrimitive<D>, min: <B as Backend>::FloatElem ) -> <B as Backend>::TensorPrimitive<D>

Clamps a tensor under a minimum value.

Arguments

tensor - The tensor to clamp.

min - The minimum value.

Returns

The clamped tensor.

fn clamp_max<const D: usize>( tensor: <B as Backend>::TensorPrimitive<D>, max: <B as Backend>::FloatElem ) -> <B as Backend>::TensorPrimitive<D>

Clamps a tensor over a maximum value.

Arguments

tensor - The tensor to clamp.

max - The maximum value.

Returns

The clamped tensor.

fn clamp<const D: usize>( tensor: <B as Backend>::TensorPrimitive<D>, min: <B as Backend>::FloatElem, max: <B as Backend>::FloatElem ) -> <B as Backend>::TensorPrimitive<D>

Clamps a tensor between a minimum and maximum value.

Arguments

tensor - The tensor to clamp.

min - The minimum value.

max - The maximum value.

Returns

The clamped tensor.

fn neg<const D: usize>( tensor: <B as Backend>::TensorPrimitive<D> ) -> <B as Backend>::TensorPrimitive<D>

Negates a tensor element-wise.

fn transpose<const D: usize>( tensor: <B as Backend>::TensorPrimitive<D> ) -> <B as Backend>::TensorPrimitive<D>

Transposes a tensor.

Arguments

tensor - The tensor to transpose.

Returns

The transposed tensor.

fn detach<const D: usize>( tensor: <B as Backend>::TensorPrimitive<D> ) -> <B as Backend>::TensorPrimitive<D>

Detaches a tensor from the computation graph.

fn set_require_grad<const D: usize>( tensor: <B as Backend>::TensorPrimitive<D>, _require_grad: bool ) -> <B as Backend>::TensorPrimitive<D>

Sets the require_grad flag of a tensor.

fn is_require_grad<const D: usize>( _tensor: &<B as Backend>::TensorPrimitive<D> ) -> bool

Returns the require_grad flag of a tensor.

fn mean<const D: usize>( tensor: <B as Backend>::TensorPrimitive<D> ) -> <B as Backend>::TensorPrimitive<1>

Mean of all elements in a tensor.

Arguments

tensor - The tensor to mean.

Returns

A scalar tensor with the mean of all elements in tensor.

fn max<const D: usize>( tensor: <B as Backend>::TensorPrimitive<D> ) -> <B as Backend>::TensorPrimitive<1>

Gets the maximum element of a tensor.

Arguments

tensor - The tensor to get the maximum elements of.

Returns

A tensor with the maximum element of tensor.

fn max_dim<const D: usize>( tensor: <B as Backend>::TensorPrimitive<D>, dim: usize ) -> <B as Backend>::TensorPrimitive<D>

Gets the maximum elements of a tensor along an axis.

Arguments

tensor - The tensor to get the maximum elements of.

dim - The dimension along which to get the maximum elements.

Returns

A tensor with the maximum elements of tensor along dim.

fn max_dim_with_indices<const D: usize>( tensor: <B as Backend>::TensorPrimitive<D>, dim: usize ) -> (<B as Backend>::TensorPrimitive<D>, <B as Backend>::IntTensorPrimitive<D>)

Gets the maximum elements of a tensor along an axis and their indices.

Arguments

tensor - The tensor to get the maximum elements of.

dim - The dimension along which to get the maximum elements.

Returns

A tuple with the maximum elements of tensor along dim and their indices.

fn min<const D: usize>( tensor: <B as Backend>::TensorPrimitive<D> ) -> <B as Backend>::TensorPrimitive<1>

Gets the minimum element of a tensor.

Arguments

tensor - The tensor to get the minimum elements of.

Returns

A tensor with the minimum element of tensor.

fn min_dim<const D: usize>( tensor: <B as Backend>::TensorPrimitive<D>, dim: usize ) -> <B as Backend>::TensorPrimitive<D>

Gets the minimum elements of a tensor along an axis.

Arguments

tensor - The tensor to get the minimum elements of.

dim - The dimension along which to get the minimum elements.

Returns

A tensor with the minimum elements of tensor along dim.

fn min_dim_with_indices<const D: usize>( tensor: <B as Backend>::TensorPrimitive<D>, dim: usize ) -> (<B as Backend>::TensorPrimitive<D>, <B as Backend>::IntTensorPrimitive<D>)

Gets the minimum elements of a tensor along an axis and their indices.

Arguments

tensor - The tensor to get the minimum elements of.

dim - The dimension along which to get the minimum elements.

Returns

A tuple with the minimum elements of tensor along dim and their indices.