burn_core::tensor::ops

Trait TransactionOps

pub trait TransactionOps<B>
where
    B: Backend,
{
    // Provided method
    fn tr_execute(
        transaction: TransactionPrimitive<B>,
    ) -> impl Future<Output = TransactionPrimitiveResult> + Send + 'static { ... }
}

Operations that are sync by nature and that can be batch together in transactions to improve compute utilization with efficient laziness.

Provided Methods

fn tr_execute( transaction: TransactionPrimitive<B>, ) -> impl Future<Output = TransactionPrimitiveResult> + Send + 'static

Executes a transaction and return its result.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.