Module burn_core::data::dataloader

Dataloader module.

Modules

batcher

Module for batching items.

Structs

BatchDataLoader

A data loader that can be used to iterate over a dataset in batches.

DataLoaderBuilder

A builder for data loaders.

DatasetIterator

Dataset iterator.

FixBatchStrategy

A strategy to batch items with a fixed batch size.

MultiThreadDataLoader

A multi-threaded data loader that can be used to iterate over a dataset.

Progress

A progress struct that can be used to track the progress of a data loader.

Enums

Message

A message that can be sent between threads.

Traits

BatchStrategy

A strategy to batch items.

DataLoader

A data loader that can be used to iterate over a dataset.

DataLoaderIterator

A data loader iterator that can be used to iterate over a data loader.

Dataset

The dataset trait defines a basic collection of items with a predefined size.