burn_core::record

Trait Record

pub trait Record<B: Backend>: Send {
    type Item<S: PrecisionSettings>: Serialize + DeserializeOwned;

    // Required methods
    fn into_item<S: PrecisionSettings>(self) -> Self::Item<S>;
    fn from_item<S: PrecisionSettings>(
        item: Self::Item<S>,
        device: &B::Device,
    ) -> Self;
}

Trait to define a family of types which can be recorded using any settings.

Required Associated Types

type Item<S: PrecisionSettings>: Serialize + DeserializeOwned

Type of the item that can be serialized and deserialized.

Required Methods

fn into_item<S: PrecisionSettings>(self) -> Self::Item<S>

Convert the current record into the corresponding item that follows the given settings.

fn from_item<S: PrecisionSettings>( item: Self::Item<S>, device: &B::Device, ) -> Self

Convert the given item into a record.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.