Trait burn_core::record::Recorder

pub trait Recorder: Send + Sync + Default + Debug + Clone {
    type Settings: PrecisionSettings;
    type RecordArgs: Clone;
    type RecordOutput;
    type LoadArgs: Clone;

    // Required methods
    fn save_item<I: Serialize>(
        &self,
        item: I,
        args: Self::RecordArgs
    ) -> Result<Self::RecordOutput, RecorderError>;
    fn load_item<I: DeserializeOwned>(
        &self,
        args: Self::LoadArgs
    ) -> Result<I, RecorderError>;

    // Provided methods
    fn record<R: Record>(
        &self,
        record: R,
        args: Self::RecordArgs
    ) -> Result<Self::RecordOutput, RecorderError> { ... }
    fn load<R: Record>(&self, args: Self::LoadArgs) -> Result<R, RecorderError> { ... }
}

Record any item implementing Serialize and DeserializeOwned.

Required Associated Types

type Settings: PrecisionSettings

Type of the settings used by the recorder.

type RecordArgs: Clone

Arguments used to record objects.

type RecordOutput

Record output type.

type LoadArgs: Clone

Arguments used to load recorded objects.

Required Methods

fn save_item<I: Serialize>( &self, item: I, args: Self::RecordArgs ) -> Result<Self::RecordOutput, RecorderError>

Saves an item.

This method is used by record to save the item.

Arguments

item - Item to save.

args - Arguments to use to save the item.

Returns

The output of the save operation.

fn load_item<I: DeserializeOwned>( &self, args: Self::LoadArgs ) -> Result<I, RecorderError>

Loads an item.

This method is used by load to load the item.

Arguments

args - Arguments to use to load the item.

Returns

The loaded item.

Provided Methods

fn record<R: Record>( &self, record: R, args: Self::RecordArgs ) -> Result<Self::RecordOutput, RecorderError>

Records an item.

Arguments

record - The item to record.

args - Arguments used to record the item.

Returns

The output of the recording.

fn load<R: Record>(&self, args: Self::LoadArgs) -> Result<R, RecorderError>

Load an item from the given arguments.